Project Files/requirements.txt
streamlit
google.generativeai
google.genai
dotenv
fpdf